+++ src/test/render/render.tsx
@@ -1,0 +1,13 @@
+import { Block } from "@tarojs/components";
+import React from "react";
+import Divider from "@Components/divider/__test__";
+
+const Component = () => {
+  return (
+    <Block>
+      <Divider />
+    </Block>
+  );
+};
+
+export default Component;