+++ src/test/render/render.tsx
@@ -1,11 +1,13 @@
 import { Block } from "@tarojs/components";
 import React from "react";
-import Divider from "@Components/divider/__test__";
+// import Divider from "@Components/divider/__test__";
+import Empty from "@Components/empty/__test__";
 
 const Component = () => {
   return (
     <Block>
-      <Divider />
+      {/* <Divider /> */}
+      <Empty />
     </Block>
   );
 };