+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - add multiSelector component ([c6f975f](https://github.com/VFiee/github-test/commit/c6f975fe0296b124074218a29fb2150d84a0653c))
+- add picker region component ([597c925](https://github.com/VFiee/github-test/commit/597c925bd2c3a3fa4e34986394dcba7a5debe735))
 - add time&date picker component ([e7b03b4](https://github.com/VFiee/github-test/commit/e7b03b49ec37dc932afe6d7d84c146eebb907f3c))
 - test changelog ([8f3acbe](https://github.com/VFiee/github-test/commit/8f3acbef99fc0cb35ddf4ef30a4fe2424bf71326))
 - 增加 Selector 组件 ([ccab599](https://github.com/VFiee/github-test/commit/ccab5994b311a643ac3f44a56287beac94af7eb0))